+++ app/src/main/java/com/pos/leaders/leaderspossystem/DataBaseAdapter/UsedPoint.java
@@ -30,12 +30,12 @@ public class UsedPoint {
 
 
 
-    public int getUnusedPointInfo(int _custmer_id) {
+    public long getUnusedPointInfo(long _custmer_id) {
 
         Cursor cur = db.rawQuery("SELECT SUM(unUsedpoint_amount) from " +  UsedPoint_TabelName + "  where custmer_id='" + _custmer_id + "'", null);
 
         if (cur.moveToFirst()) {
-            return cur.getInt(0);
+            return cur.getLong(0);
         }
         return  0;
     }
@@ -57,17 +57,14 @@ public class UsedPoint {
     public SQLiteDatabase getDatabaseInstance() {
         return db;
     }
-    public int insertEntry(int  sale_id, int point,int custmer_id){
+    public int insertEntry(long sale_id, long point,long custmer_id){
         ContentValues val = new ContentValues();
         //Assign values for each row.
         val.put(UsedPoint_COLUMN_Sale_Id,sale_id);
         val.put(UsedPoint_COLUMN_Point,point);
-
         val.put(UsedPoint_COLUMN_Custmer,custmer_id);
 
-
         try {
-
             db.insert( UsedPoint_TabelName, null, val);
             return 1;
         } catch (SQLException ex) {
