+++ app/src/main/java/com/pos/leaders/leaderspossystem/HomeActivity.java
@@ -99,7 +99,7 @@ public class HomeActivity extends Activity {
 					intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
 					intent.putExtra(LogInActivity.LEADPOS_MAKE_A_REPORT, str);
 					intent.putExtra("permissions_name",u.getPermtionName());
-					int scheduleID = scheduleWorkersDBAdapter.insertEntry(u.getId());
+					long scheduleID = scheduleWorkersDBAdapter.insertEntry(u.getId());
 					SESSION._SCHEDULEWORKERS = new ScheduleWorkers(scheduleID, u.getId(), new Date(), new Date());
 					/*
 					String message = "hello";
