+++ app/src/main/java/com/pos/leaders/leaderspossystem/LogInActivity.java
@@ -143,7 +143,7 @@ public class LogInActivity extends Activity implements View.OnClickListener {
                 intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
                 intent.putExtra(LogInActivity.LEADPOS_MAKE_A_REPORT, LogInActivity.LEADPOS_MAKE_A_REPORT);
 
-                int scheduleID = scheduleWorkersDBAdapter.insertEntry(user.getId());
+                long scheduleID = scheduleWorkersDBAdapter.insertEntry(user.getId());
                 SESSION._SCHEDULEWORKERS = new ScheduleWorkers(scheduleID, user.getId(), new Date(), new Date());
                 userDBAdapter.close();
                 scheduleWorkersDBAdapter.close();
